Completed upgrading "unsigned" to "unsigned long" as the integer fact type everywhere.

+++ facts.c
@@ -18,8 +18,8 @@
 #include "facts.h"
 
 const char* fact_username = 0;
-unsigned fact_userid = 0;
-unsigned fact_groupid = 0;
+unsigned long fact_userid = 0;
+unsigned long fact_groupid = 0;
 const char* fact_realname = 0;
 const char* fact_directory = 0;
 const char* fact_shell = 0;
